refactor(api): move division usecase to its bounded context
