+++ api/src/prescription/campaign/domain/usecases/index.js
@@ -29,6 +29,7 @@ import * as campaignParticipationRepository from '../../infrastructure/repositor
 import * as campaignProfilesCollectionParticipationSummaryRepository from '../../infrastructure/repositories/campaign-profiles-collection-participation-summary-repository.js';
 import * as campaignReportRepository from '../../infrastructure/repositories/campaign-report-repository.js';
 import * as campaignToJoinRepository from '../../infrastructure/repositories/campaign-to-join-repository.js';
+import * as divisionRepository from '../../infrastructure/repositories/division-repository.js';
 import * as targetProfileRepository from '../../infrastructure/repositories/target-profile-repository.js';
 import * as campaignCsvExportService from '../services/campaign-csv-export-service.js';
 import * as campaignUpdateValidator from '../validators/campaign-update-validator.js';
@@ -50,6 +51,7 @@ const dependencies = {
   codeGenerator,
   campaignUpdateValidator,
   competenceRepository,
+  divisionRepository,
   knowledgeElementSnapshotRepository,
   knowledgeElementRepository,
   learningContentRepository,
